Merge branch 'feature/GUI-AssignmentViewing' into dev

+++ Domain/src/main/java/javaminor/al/entities/concrete/MaintenanceWork.java
@@ -26,7 +26,7 @@ public class MaintenanceWork extends PersistentEntity {
     @Temporal(TemporalType.TIMESTAMP)
     private Calendar finishedOn;
 
-    private int workHours;
+    private double workHours;
 
     private String comments;
